feat(providers)!: add CancellationToken and widen id collections to int64
Adds `CancellationToken cancellationToken = default` as a trailing parameter to
all 73 provider methods and forwards it through all 116 adapter call sites.
Nothing could previously be cancelled, so a hosted service could not shut down
cleanly mid-sync. Adding this later would break the ABI again, so it belongs in
the same major.

Also widens the group and member id collections from int to long:

  IEmmaGroupProvider.AddMembersToGroup / RemoveMembersFromGroup
  IEmmaMemberProvider.AddMemberToGroups / RemoveMemberFromGroups
  AddMember.GroupIds, AddMembers.GroupIds,
  RemoveMemberGroups.GroupIds, SignupMember.GroupIds

This closes kylegregory/EmmaSharp#48, open since 2019. Worth noting the request
side was already correct -- MemberIdList.MemberIds has been List<long> for years
-- so only the returned member ids overflowed, exactly as that issue reported.

BREAKING CHANGE: provider methods take a trailing CancellationToken; group and
member id collections are long rather than int.

+++ src/EmmaSharper/Models/Members/AddMember.cs
@@ -9,7 +9,7 @@ namespace EmmaSharper
         /// <inheritdoc cref="object.Object"/>
         public AddMember()
         {
-            GroupIds = new List<int>();
+            GroupIds = new List<long>();
         }
 
         /// <summary>
@@ -28,7 +28,7 @@ namespace EmmaSharper
         /// Optional. Add imported members to this list of groups.
         /// </summary>
         [JsonPropertyName("group_ids")]
-        public List<int> GroupIds { get; set; }
+        public List<long> GroupIds { get; set; }
 
         /// <summary>
         /// Optional. Fires related field change auto-responders when set to true.
